chore: separate treeshaking and snapshot different rollup tests
