+++ cspell.json
@@ -23,7 +23,7 @@
     "/CHANGELOG.md",
     "/packages/rolldown/src/rollup.d.ts",
     "/crates/rolldown/tests/esbuild/splitting/ignore-test.md",
-    "/packages/rollup-tests/src/ignored-by-supported-features.md",
+    "/packages/rollup-tests/src/ignored-by-unsupported-features.md",
     "crates/rolldown/tests/esbuild/**/*",
     "/crates/rolldown_utils/src/light_guess.rs",
     "/pnpm-lock.yaml",
