+++ packages/rollup-tests/src/status.json
@@ -1,7 +1,9 @@
 {
   "failed": 0,
   "skipFailed": 0,
-  "ignored": 434,
-  "ignoredByUnsupportedFeatures": 402,
+  "ignored": 38,
+  "ignored(unsupported features)": 402,
+  "ignored(treeshaking)": 285,
+  "ignored(behavior passed, snapshot different)": 113,
   "passed": 680
 }
