+++ packages/rollup-tests/src/status.md
@@ -2,6 +2,8 @@
 |----| ---- |
 | failed | 0|
 | skipFailed | 0|
-| ignored | 434|
-| ignoredByUnsupportedFeatures | 402|
+| ignored | 38|
+| ignored(unsupported features) | 402|
+| ignored(treeshaking) | 285|
+| ignored(behavior passed, snapshot different) | 113|
 | passed | 680|
